ProductMicro authentication template has been put in place.
